+++ src/main/java/com/salessparrow/api/dto/formatter/GetTaskDetailsFormatterDto.java
@@ -1,0 +1,15 @@
+package com.salessparrow.api.dto.formatter;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import com.salessparrow.api.dto.entities.TaskEntity;
+
+import lombok.Data;
+
+@Data
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class GetTaskDetailsFormatterDto {
+
+	private TaskEntity taskDetail;
+
+}