+++ src/main/java/com/salessparrow/api/lib/crmActions/getAccountTaskDetails/GetAccountTaskDetails.java
@@ -1,0 +1,16 @@
+package com.salessparrow.api.lib.crmActions.getAccountTaskDetails;
+
+import org.springframework.stereotype.Component;
+
+import com.salessparrow.api.domain.User;
+import com.salessparrow.api.dto.formatter.GetTaskDetailsFormatterDto;
+
+/**
+ * GetAccountTaskDetails is an interface for the getTaskDetails action for the CRM.
+ */
+@Component
+public interface GetAccountTaskDetails {
+
+	public GetTaskDetailsFormatterDto getTaskDetails(User user, String taskId);
+
+}